*** empty log message ***
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@1520 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/src/gtk/iupgtk_list.c
@@ -550,8 +550,8 @@ static int gtkListSetPaddingAttrib(Ihandle* ih, const char* value)
 #if GTK_CHECK_VERSION(2, 10, 0)
     GtkEntry* entry;
     GtkBorder border;
-    border.bottom = border.top = ih->data->vert_padding;
-    border.left = border.right = ih->data->horiz_padding;
+    border.bottom = border.top = (gint16)ih->data->vert_padding;
+    border.left = border.right = (gint16)ih->data->horiz_padding;
     entry = (GtkEntry*)iupAttribGet(ih, "_IUPGTK_ENTRY");
     gtk_entry_set_inner_border(entry, &border);
 #endif
